fix fp32 activation leak from RecurrentBlock into coda under bf16
Under FSDP MixedPrecision(bf16), `h` enters RecurrentBlock.forward as bf16.
Three spots inside the loop silently promoted activations back to fp32:

  - `cumulative_p = torch.zeros(B, T, device=h.device)` — no dtype arg,
    so PyTorch picks fp32 regardless of `h.dtype`.
  - `still_running.float()` (x2) — hard-codes fp32 even when h is bf16.

`cumulative_p + p * still_running.float()` then yields fp32, and via
`h_out + weight.unsqueeze(-1) * h` the fp32 weight broadcasts and promotes
`h_out` itself. That fp32 output flows into the coda transformer blocks,
whose Linear layers hold bf16 weights, producing:

  RuntimeError: expected mat1 and mat2 to have the same dtype,
                but got: float != c10::BFloat16

Aligning all three to `h.dtype` keeps the recurrent output in the
activation dtype. In fp32 these are no-ops; in bf16 they prevent the leak.

+++ open_mythos/main.py
@@ -823,7 +823,11 @@ class RecurrentBlock(nn.Module):
         B, T, D = h.shape
 
         halted = torch.zeros(B, T, device=h.device, dtype=torch.bool)
-        cumulative_p = torch.zeros(B, T, device=h.device)
+        # Match h's dtype so the ACT weight accumulation into h_out stays in
+        # the activation dtype (bf16 under FSDP MixedPrecision) — a naive
+        # fp32 cumulative_p silently promotes h_out via broadcasting and
+        # feeds fp32 into the coda's bf16 Linear layers.
+        cumulative_p = torch.zeros(B, T, device=h.device, dtype=h.dtype)
         h_out = torch.zeros_like(h)
 
         for t in range(n_loops):
@@ -848,10 +852,10 @@ class RecurrentBlock(nn.Module):
                 remainder,
                 p,
             )
-            weight = weight * still_running.float()
+            weight = weight * still_running.to(h.dtype)
             h_out = h_out + weight.unsqueeze(-1) * h
 
-            cumulative_p = cumulative_p + p * still_running.float()
+            cumulative_p = cumulative_p + p * still_running.to(h.dtype)
             halted = halted | (cumulative_p >= self.cfg.act_threshold)
 
             # Only short-circuit when there is no KV cache to keep consistent.
